requirements-service.txt
numpy==1.16.3
spacy==2.1.3
transformers==2.2.2
Cython==0.29.10
tqdm==4.32.2
neuralcoref==4.0
argparse
scikit-learn
bert-extractive-summarizer==0.2.2
Flask
flask-cors
nltk